% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_module_features}
\alias{analyze_module_features}
\title{Analyze Module Features}
\usage{
analyze_module_features(modules, ppi_network)
}
\arguments{
\item{modules}{List of classified modules}

\item{ppi_network}{Original PPI network (igraph object)}
}
\value{
Data frame with module features
}
\description{
Analyze topological and functional features of modules
}
\examples{
module_analysis <- analyze_module_features(
  classified_modules,
  ppi_network = ppi_network
)
}
