% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{classify_modules}
\alias{classify_modules}
\title{Classify Modules}
\usage{
classify_modules(
  modules,
  classification = "modification_type",
  rna_proteins = NULL
)
}
\arguments{
\item{modules}{List of annotated modules}

\item{classification}{Type of classification ("modification_type" or "functional_role")}

\item{rna_proteins}{Data frame containing RNA modification proteins}
}
\value{
Classified modules with additional metadata
}
\description{
Classify modules by modification type or functional role
}
\examples{
classified_modules <- classify_modules(
  annotated_modules,
  classification = "modification_type",
  rna_proteins = rna_mod_proteins
)
}
