% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{identify_modules_iterative}
\alias{identify_modules_iterative}
\title{Iterative Module Identification with Target Proteins}
\usage{
identify_modules_iterative(
  ppi_network,
  min_module_size = 10,
  max_module_size = 1000,
  target_proteins = NULL
)
}
\arguments{
\item{ppi_network}{An igraph object representing the PPI network}

\item{min_module_size}{Minimum number of nodes in a module}

\item{max_module_size}{Maximum number of nodes in a module}

\item{target_proteins}{Character vector of target proteins to focus on}
}
\value{
A list containing all identified modules that contain target proteins
}
\description{
Identify network modules using louvain and fast_greedy algorithms iteratively,
focusing on modules that contain target proteins
}
\examples{
modules <- identify_modules_iterative(
  ppi_network = ppi_network,
  min_module_size = 10,
  max_module_size = 1000,
  target_proteins = c("DNMT1", "DNMT3A", "TET1")
)
}
