% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_module_features}
\alias{plot_module_features}
\title{Plot Module Features}
\usage{
plot_module_features(module_analysis, output_dir = "module_plots")
}
\arguments{
\item{module_analysis}{Data frame from analyze_module_features()}

\item{output_dir}{Directory to save plots}
}
\description{
Visualize module properties and characteristics
}
\examples{
plot_module_features(module_analysis, output_dir = "module_plots/")
}
