% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_module_scores}
\alias{plot_module_scores}
\title{Plot Module Scores}
\usage{
plot_module_scores(score_comparison, output_dir = "score_plots")
}
\arguments{
\item{score_comparison}{Output from compare_module_scores()}

\item{output_dir}{Directory to save plots}
}
\description{
Visualize module scores across sample groups
}
\examples{
plot_module_scores(score_comparison, output_dir = "score_plots/")
}
