% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_network_comparison}
\alias{plot_network_comparison}
\title{Plot Network Comparison}
\usage{
plot_network_comparison(
  networks_list,
  network_names = NULL,
  save_path = NULL,
  width = 12,
  height = 8
)
}
\arguments{
\item{networks_list}{List of igraph objects to compare}

\item{network_names}{Names for the networks}

\item{save_path}{Optional path to save the comparison plot}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
ggplot object
}
\description{
Compare multiple networks side by side
}
\examples{
\dontrun{
# Create sample networks
library(igraph)
net1 <- erdos.renyi.game(50, 0.1)
net2 <- erdos.renyi.game(60, 0.15)

# Compare networks
plots <- plot_network_comparison(
  networks_list = list(net1, net2),
  network_names = c("Network 1", "Network 2")
)

# Display comparison
print(plots$density)
}

}
