% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_network_summary}
\alias{plot_network_summary}
\title{Plot Network Summary}
\usage{
plot_network_summary(network, save_path = NULL, width = 8, height = 6)
}
\arguments{
\item{network}{An igraph object representing the PPI network}

\item{save_path}{Optional path to save the plots}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
List of ggplot objects
}
\description{
Create summary plots for network analysis
}
\examples{
\dontrun{
# Create a sample network
library(igraph)
g <- erdos.renyi.game(50, 0.1)
V(g)$name <- paste0("GENE", 1:50)

# Generate summary plots
plots <- plot_network_summary(g)

# Display degree distribution
print(plots$degree_distribution)
}

}
